Refactor, moved code from model to service
